Moved Elasticsearch tests as integration tests
