+++ logisland-components/logisland-services/logisland-service-elasticsearch/logisland-service-elasticsearch_2_4_0-client/src/integration-test/java/com/hurence/logisland/service/elasticsearch/ESRule.java
