+++ logisland-components/logisland-services/logisland-service-elasticsearch/logisland-service-elasticsearch_6_6_2-client/src/integration-test/java/com/hurence/logisland/service/elasticsearch/Elasticsearch_6_6_2_ClientServiceIT.java
@@ -1,18 +1,19 @@
-/**
+ /**
  * Copyright (C) 2016 Hurence (support@hurence.com)
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.hurence.logisland.service.elasticsearch;
 
 import com.hurence.logisland.classloading.PluginProxy;
@@ -20,15 +21,20 @@ import com.hurence.logisland.component.InitializationException;
 import com.hurence.logisland.component.PropertyDescriptor;
 import com.hurence.logisland.controller.ControllerServiceInitializationContext;
 import com.hurence.logisland.processor.ProcessException;
+import com.hurence.logisland.record.FieldType;
+import com.hurence.logisland.record.Record;
+import com.hurence.logisland.record.StandardRecord;
 import com.hurence.logisland.service.datastore.InvalidMultiGetQueryRecordException;
 import com.hurence.logisland.service.datastore.MultiGetQueryRecord;
 import com.hurence.logisland.service.datastore.MultiGetResponseRecord;
 import com.hurence.logisland.util.runner.TestRunner;
 import com.hurence.logisland.util.runner.TestRunners;
+import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.bulk.BulkItemResponse;
 import org.elasticsearch.action.bulk.BulkProcessor;
 import org.elasticsearch.action.bulk.BulkRequest;
 import org.elasticsearch.action.bulk.BulkResponse;
+import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.common.unit.ByteSizeUnit;
 import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.common.unit.TimeValue;
@@ -40,21 +46,22 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.*;
+import java.util.function.BiConsumer;
 
-public class TestElasticsearch_2_4_0_ClientService {
+public class Elasticsearch_6_6_2_ClientServiceIT {
 
-    private static final String MAPPING1 = "{'properties':{'name':{'type': 'string', 'index': 'not_analyzed'},'val':{'type':'integer'}}}";
-    private static final String MAPPING2 = "{'properties':{'name':{'type': 'string', 'index': 'not_analyzed'},'val':{'type': 'string', 'index': 'not_analyzed'}}}";
+    private static final String MAPPING1 = "{'properties':{'name':{'type': 'text'},'val':{'type':'integer'}}}";
+    private static final String MAPPING2 = "{'properties':{'name':{'type': 'text'},'val':{'type': 'text'}}}";
     private static final String MAPPING3 =
-            "{'dynamic':'strict','properties':{'name':{'type': 'string', 'index': 'not_analyzed'},'xyz':{'type': 'string', 'index': 'not_analyzed'}}}";
+            "{'dynamic':'strict','properties':{'name':{'type': 'text'},'xyz':{'type': 'text'}}}";
 
-    private static Logger logger = LoggerFactory.getLogger(TestElasticsearch_2_4_0_ClientService.class);
+    private static Logger logger = LoggerFactory.getLogger(Elasticsearch_6_6_2_ClientServiceIT.class);
 
     @Rule
     public final ESRule esRule = new ESRule();
 
 
-    private class MockElasticsearchClientService extends Elasticsearch_2_4_0_ClientService {
+    private class MockElasticsearchClientService extends Elasticsearch_6_6_2_ClientService {
 
         @Override
         protected void createElasticsearchClient(ControllerServiceInitializationContext context) throws ProcessException {
@@ -65,17 +72,15 @@ public class TestElasticsearch_2_4_0_ClientService {
         }
 
         @Override
-        protected void createBulkProcessor(ControllerServiceInitializationContext context)
-        {
+        protected void createBulkProcessor(ControllerServiceInitializationContext context) {
+
             if (bulkProcessor != null) {
                 return;
             }
 
-            /**
-             * create the bulk processor
-             */
-            bulkProcessor = BulkProcessor.builder(
-                    esClient,
+            // create the bulk processor
+
+            BulkProcessor.Listener listener =
                     new BulkProcessor.Listener() {
                         @Override
                         public void beforeBulk(long l, BulkRequest bulkRequest) {
@@ -88,7 +93,7 @@ public class TestElasticsearch_2_4_0_ClientService {
                             if (bulkResponse.hasFailures()) {
                                 getLogger().warn("There was failures while executing bulk [id:{}]," +
                                                 " done bulk request in {} ms with failure = {}",
-                                        new Object[]{l, bulkResponse.getTookInMillis(), bulkResponse.buildFailureMessage()});
+                                        new Object[]{l, bulkResponse.getTook().getMillis(), bulkResponse.buildFailureMessage()});
                                 for (BulkItemResponse item : bulkResponse.getItems()) {
                                     if (item.isFailed()) {
                                         errors.put(item.getId(), item.getFailureMessage());
@@ -102,13 +107,18 @@ public class TestElasticsearch_2_4_0_ClientService {
                             getLogger().error("something went wrong while bulk loading events to es : {}", new Object[]{throwable.getMessage()});
                         }
 
-                    })
+                    };
+
+            BiConsumer<BulkRequest, ActionListener<BulkResponse>> bulkConsumer =
+                    (request, bulkListener) -> esClient.bulkAsync(request, RequestOptions.DEFAULT, bulkListener);
+            bulkProcessor = BulkProcessor.builder(bulkConsumer, listener)
                     .setBulkActions(1000)
                     .setBulkSize(new ByteSizeValue(10, ByteSizeUnit.MB))
                     .setFlushInterval(TimeValue.timeValueSeconds(1))
                     .setConcurrentRequests(2)
                     //.setBackoffPolicy(getBackOffPolicy(context))
                     .build();
+
         }
 
         @Override
@@ -121,8 +131,7 @@ public class TestElasticsearch_2_4_0_ClientService {
 
     }
 
-    private ElasticsearchClientService configureElasticsearchClientService(final TestRunner runner) throws InitializationException
-    {
+    private ElasticsearchClientService configureElasticsearchClientService(final TestRunner runner) throws InitializationException {
         final MockElasticsearchClientService elasticsearchClientService = new MockElasticsearchClientService();
 
         runner.addControllerService("elasticsearchClient", elasticsearchClientService);
@@ -158,24 +167,20 @@ public class TestElasticsearch_2_4_0_ClientService {
         Assert.assertEquals(true, elasticsearchClientService.existsCollection("foo"));
 
         // Define another index
-        elasticsearchClientService.createCollection("bar",2, 1);
+        elasticsearchClientService.createCollection("bar", 2, 1);
         Assert.assertEquals(true, elasticsearchClientService.existsCollection("foo"));
 
         // Add a mapping to foo
         result = elasticsearchClientService.putMapping("foo", "type1", MAPPING1.replace('\'', '"'));
-//        Assert.assertEquals(true, result);
+        Assert.assertEquals(true, result);
 
         // Add the same mapping again
         result = elasticsearchClientService.putMapping("foo", "type1", MAPPING1.replace('\'', '"'));
-        //       Assert.assertEquals(true, result);
-
-        // Add the same mapping again under a different doctype
-        result = elasticsearchClientService.putMapping("foo", "type2", MAPPING1.replace('\'', '"'));
-        //       Assert.assertEquals(true, result);
+        Assert.assertEquals(true, result);
 
         // Update a mapping with an incompatible mapping -- should fail
-        //result = elasticsearchClientService.putMapping("foo", "type2", MAPPING2.replace('\'', '"'));
-        //Assert.assertEquals(false, result);
+        // result = elasticsearchClientService.putMapping("foo", "type2", MAPPING2.replace('\'', '"'));
+        // Assert.assertEquals(false, result);
 
         // create alias
         elasticsearchClientService.createAlias("foo", "aliasFoo");
@@ -206,7 +211,7 @@ public class TestElasticsearch_2_4_0_ClientService {
         // The document (doc1) of type "type1" already in index "foo" cannot be inserted into index "baz" as type1
         // because that means applying its source to MAPPING3 - but MAPPING3 is strict and does not define property
         // "val", so the insert fails.
-        elasticsearchClientService.createCollection("baz", 2, 1);
+        elasticsearchClientService.createCollection("baz",2, 1);
         elasticsearchClientService.putMapping("baz", "type1", MAPPING3.replace('\'', '"'));
 
       /*  try {
@@ -248,7 +253,7 @@ public class TestElasticsearch_2_4_0_ClientService {
         Assert.assertEquals(false, elasticsearchClientService.existsCollection(index));
 
         // Create the index
-        elasticsearchClientService.createCollection(index, 2, 1);
+        elasticsearchClientService.createCollection(index,2, 1);
         Assert.assertEquals(true, elasticsearchClientService.existsCollection(index));
 
         // Put a document in the bulk processor :
@@ -293,13 +298,76 @@ public class TestElasticsearch_2_4_0_ClientService {
 
     }
 
+
+    @Test
+    public void testBulkPutGeopoint() throws InitializationException, IOException, InterruptedException {
+        final String index = "future_factory";
+        final String type = "factory";
+        final String docId = "modane_factory";
+        Record record = new StandardRecord("factory")
+                .setId(docId)
+                .setStringField("address", "rue du Frejus")
+                .setField("latitude", FieldType.FLOAT, 45.4f)
+                .setField("longitude", FieldType.FLOAT, 45.4f);
+
+        final TestRunner runner = TestRunners.newTestRunner(new TestProcessor());
+
+        // create the controller service and link it to the test processor :
+        final ElasticsearchClientService elasticsearchClientService = configureElasticsearchClientService(runner);
+
+        // Verify the index does not exist
+        Assert.assertEquals(false, elasticsearchClientService.existsCollection(index));
+
+        // Create the index
+        elasticsearchClientService.createCollection(index, 2, 1);
+        Assert.assertEquals(true, elasticsearchClientService.existsCollection(index));
+
+        // Put a document in the bulk processor :
+        String document1 = ElasticsearchRecordConverter.convertToString(record);
+        elasticsearchClientService.bulkPut(index, type, document1, Optional.of(docId));
+        // Flush the bulk processor :
+        elasticsearchClientService.bulkFlush();
+        Thread.sleep(2000);
+        try {
+            // Refresh the index :
+            elasticsearchClientService.refreshCollection(index);
+        } catch (Exception e) {
+            logger.error("Error while refreshing the index : " + e.toString());
+        }
+
+        long documentsNumber = 0;
+
+        try {
+            documentsNumber = elasticsearchClientService.countCollection(index);
+        } catch (Exception e) {
+            logger.error("Error while counting the number of documents in the index : " + e.toString());
+        }
+
+        Assert.assertEquals(1, documentsNumber);
+
+        List<MultiGetQueryRecord> multiGetQueryRecords = new ArrayList<>();
+        ArrayList<String> documentIds = new ArrayList<>();
+        List<MultiGetResponseRecord> multiGetResponseRecords = new ArrayList<>();
+
+
+        // Make sure a dummy query returns no result :
+        documentIds.add(docId);
+        try {
+            multiGetQueryRecords.add(new MultiGetQueryRecord(index, type, new String[]{"location", "id"}, new String[]{}, documentIds));
+        } catch (InvalidMultiGetQueryRecordException e) {
+            e.printStackTrace();
+        }
+        multiGetResponseRecords = elasticsearchClientService.multiGet(multiGetQueryRecords);
+        Assert.assertEquals(1, multiGetResponseRecords.size()); // number of documents retrieved
+
+    }
+
+
     @Test
     public void testMultiGet() throws InitializationException, IOException, InterruptedException, InvalidMultiGetQueryRecordException {
         final String index1 = "index1";
         final String index2 = "index2";
         final String type1 = "type1";
-        final String type2 = "type2";
-        final String type3 = "type3";
 
         Map<String, Object> document1 = new HashMap<>();
         final String docId1 = "id1";
@@ -346,8 +414,8 @@ public class TestElasticsearch_2_4_0_ClientService {
         elasticsearchClientService.bulkPut(index1, type1, document2, Optional.of(docId2));
         elasticsearchClientService.bulkPut(index1, type1, document3, Optional.of(docId3));
         elasticsearchClientService.bulkPut(index2, type1, document1, Optional.of(docId1));
-        elasticsearchClientService.bulkPut(index2, type2, document2, Optional.of(docId2));
-        elasticsearchClientService.bulkPut(index2, type3, document3, Optional.of(docId3));
+        elasticsearchClientService.bulkPut(index2, type1, document2, Optional.of(docId2));
+        elasticsearchClientService.bulkPut(index2, type1, document3, Optional.of(docId3));
         // Flush the bulk processor :
         elasticsearchClientService.bulkFlush();
         Thread.sleep(2000);
@@ -373,13 +441,13 @@ public class TestElasticsearch_2_4_0_ClientService {
         List<MultiGetQueryRecord> multiGetQueryRecords = new ArrayList<>();
         ArrayList<String> documentIds = new ArrayList<>();
         ArrayList<String> documentIds_2 = new ArrayList<>();
-        List<MultiGetResponseRecord> multiGetResponseRecords = new ArrayList<>();
+        List<MultiGetResponseRecord> multiGetResponseRecords;
         String[] fieldsToInclude = {"field_b*", "field*1"};
         String[] fieldsToExclude = {"field_*2"};
 
         // Make sure a dummy query returns no result :
         documentIds.add(docId1);
-        multiGetQueryRecords.add(new MultiGetQueryRecord("dummy", "",new String[]{"dummy"},new String[]{}, documentIds));
+        multiGetQueryRecords.add(new MultiGetQueryRecord("dummy", "", new String[]{"dummy"}, new String[]{}, documentIds));
         multiGetResponseRecords = elasticsearchClientService.multiGet(multiGetQueryRecords);
         Assert.assertEquals(0, multiGetResponseRecords.size()); // number of documents retrieved
 
@@ -397,7 +465,7 @@ public class TestElasticsearch_2_4_0_ClientService {
         Assert.assertEquals(type1, multiGetResponseRecords.get(0).getTypeName());
         Assert.assertEquals(docId1, multiGetResponseRecords.get(0).getDocumentId());
         Assert.assertEquals(5, multiGetResponseRecords.get(0).getRetrievedFields().size()); // number of fields retrieved for the document
-        multiGetResponseRecords.get(0).getRetrievedFields().forEach((k,v) -> document1.get(k).equals(v.toString()));
+        multiGetResponseRecords.get(0).getRetrievedFields().forEach((k, v) -> document1.get(k).equals(v.toString()));
 
         multiGetQueryRecords.clear();
         documentIds.clear();
@@ -414,22 +482,22 @@ public class TestElasticsearch_2_4_0_ClientService {
         multiGetResponseRecords.forEach(responseRecord -> Assert.assertEquals(index1, responseRecord.getCollectionName())); // verify that all retrieved are in index1
         multiGetResponseRecords.forEach(responseRecord -> Assert.assertEquals(type1, responseRecord.getTypeName())); // verify that the type of all retrieved documents is type1
         multiGetResponseRecords.forEach(responseRecord -> {
-                if (responseRecord.getDocumentId() == docId1) {
-                    Assert.assertEquals(3, responseRecord.getRetrievedFields().size()); // for document1, verify that 3 fields has been retrieved
-                    // verify that the 3 retrieved fields are the correct ones :
-                    Assert.assertEquals(true, responseRecord.getRetrievedFields().containsKey("field_beg_1"));
-                    Assert.assertEquals(true, responseRecord.getRetrievedFields().containsKey("field_beg_3"));
-                    Assert.assertEquals(true, responseRecord.getRetrievedFields().containsKey("field_fin_1"));
-                    // verify that the values of the 3 retrieved fields are the correct ones :
-                    Assert.assertEquals("field_beg_1_document1_value", responseRecord.getRetrievedFields().get("field_beg_1").toString());
-                    Assert.assertEquals("field_beg_3_document1_value", responseRecord.getRetrievedFields().get("field_beg_3").toString());
-                    Assert.assertEquals("field_fin_1_document1_value", responseRecord.getRetrievedFields().get("field_fin_1").toString());
-                }
-                if (responseRecord.getDocumentId() == docId2)
-                    Assert.assertEquals(3, responseRecord.getRetrievedFields().size()); // for document2, verify that 3 fields has been retrieved
-                if (responseRecord.getDocumentId() == docId3)
-                    Assert.assertEquals(2, responseRecord.getRetrievedFields().size()); // for document3, verify that 2 fields has been retrieved
-            });
+            if (responseRecord.getDocumentId() == docId1) {
+                Assert.assertEquals(3, responseRecord.getRetrievedFields().size()); // for document1, verify that 3 fields has been retrieved
+                // verify that the 3 retrieved fields are the correct ones :
+                Assert.assertEquals(true, responseRecord.getRetrievedFields().containsKey("field_beg_1"));
+                Assert.assertEquals(true, responseRecord.getRetrievedFields().containsKey("field_beg_3"));
+                Assert.assertEquals(true, responseRecord.getRetrievedFields().containsKey("field_fin_1"));
+                // verify that the values of the 3 retrieved fields are the correct ones :
+                Assert.assertEquals("field_beg_1_document1_value", responseRecord.getRetrievedFields().get("field_beg_1").toString());
+                Assert.assertEquals("field_beg_3_document1_value", responseRecord.getRetrievedFields().get("field_beg_3").toString());
+                Assert.assertEquals("field_fin_1_document1_value", responseRecord.getRetrievedFields().get("field_fin_1").toString());
+            }
+            if (responseRecord.getDocumentId() == docId2)
+                Assert.assertEquals(3, responseRecord.getRetrievedFields().size()); // for document2, verify that 3 fields has been retrieved
+            if (responseRecord.getDocumentId() == docId3)
+                Assert.assertEquals(2, responseRecord.getRetrievedFields().size()); // for document3, verify that 2 fields has been retrieved
+        });
 
         multiGetQueryRecords.clear();
         documentIds.clear();
@@ -444,7 +512,7 @@ public class TestElasticsearch_2_4_0_ClientService {
         documentIds_2.add(docId1);
         documentIds_2.add(docId1);
         documentIds_2.add(docId1);
-        multiGetQueryRecords.add(new MultiGetQueryRecord(index2, null , fieldsToInclude, null, documentIds_2));
+        multiGetQueryRecords.add(new MultiGetQueryRecord(index2, null, fieldsToInclude, null, documentIds_2));
         multiGetResponseRecords = elasticsearchClientService.multiGet(multiGetQueryRecords);
 
         Assert.assertEquals(5, multiGetResponseRecords.size()); // verify that 5 documents has been retrieved
@@ -466,31 +534,31 @@ public class TestElasticsearch_2_4_0_ClientService {
 
         List<MultiGetQueryRecord> multiGetQueryRecords = new ArrayList<>();
 
-        String errorMessage ="";
+        String errorMessage = "";
 
         // Validate null index behaviour :
         try {
             multiGetQueryRecords.add(new MultiGetQueryRecord(null, null, null, null, null));
-        }  catch (InvalidMultiGetQueryRecordException e) {
+        } catch (InvalidMultiGetQueryRecordException e) {
             errorMessage = e.getMessage();
         }
-        Assert.assertEquals(errorMessage,"The index name cannot be null");
+        Assert.assertEquals(errorMessage, "The index name cannot be null");
 
         // Validate empty index behaviour :
         try {
             multiGetQueryRecords.add(new MultiGetQueryRecord("", null, null, null, null));
-        }  catch (InvalidMultiGetQueryRecordException e) {
+        } catch (InvalidMultiGetQueryRecordException e) {
             errorMessage = e.getMessage();
         }
-        Assert.assertEquals(errorMessage,"The index name cannot be empty");
+        Assert.assertEquals(errorMessage, "The index name cannot be empty");
 
         // Validate null documentIds behaviour :
         try {
             multiGetQueryRecords.add(new MultiGetQueryRecord("dummy", null, null, null, null));
-        }  catch (InvalidMultiGetQueryRecordException e) {
+        } catch (InvalidMultiGetQueryRecordException e) {
             errorMessage = e.getMessage();
         }
-        Assert.assertEquals(errorMessage,"The list of document ids cannot be null");
+        Assert.assertEquals(errorMessage, "The list of document ids cannot be null");
 
         // Make sure no invalid MultiGetQueryRecord has been added to multiGetQueryRecords list :
         Assert.assertEquals(0, multiGetQueryRecords.size());
